requirements.txt
absl-py==1.4.0
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
flatbuffers==23.5.26
fonttools==4.39.4
gast==0.4.0
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.2
h5py==3.8.0
idna==3.4
joblib==1.2.0
keras==2.13.1rc0
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
sklearn==0.0.post5
tensorboard==2.13.0
tensorboard-data-server==0.7.0
tensorflow==2.13.0rc1
tensorflow-estimator==2.13.0rc0
tensorflow-macos==2.13.0rc1
termcolor==2.3.0
threadpoolctl==3.1.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.4
wrapt==1.15.0
